[Chromium] Implement PluginViewBase::getFormValue https://bugs.webkit.org/show_bug.cgi?id=64434
Reviewed by Darin Fisher.

This patch was landed as r102873, but reverted because of Chromium Mac build failure.
The fix was landed as Chromium r114920. The patch should be safe to land.

No new tests. No behavior change at this time. We should add a test for PPAPI after chromium supports the feature.

* public/WebPlugin.h: Added getFormValue().
* src/WebPluginContainerImpl.cpp:
(WebKit::WebPluginContainerImpl::getFormValue): Added.
* src/WebPluginContainerImpl.h: Added getFormValue().


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103147 268f45cc-cd09-0410-ab3c-d52691b4dbfc