Replace webkitperl/run-leaks_unittest/RunLeaks.pm with webkitperl/LoadAsModule.pm https://bugs.webkit.org/show_bug.cgi?id=74836
Reviewed by David Kilzer.

webkitperl/run-leaks_unittest/RunLeaks.pm can be used for unit-testing of run-leaks only.
This patch creates more generalized webkitperl/LoadAsModule.pm, which can be also used
for unit-testing of other Perl scripts. We are planning to use it for unit-testing
of prepare-ChangeLog.

Using LoadAsModule.pm, you can load a Perl script as follows.
The first argument is the arbitrary package name you want to use, and the second argument
is the name of the Perl script you want to load.

    use LoadAsModule qw(RunLeaks run-leaks);
    RunLeaks::someFunctionDefinedInRunLeaks(...);

Tests: Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v1.0.pl
       Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v2.0-new.pl
       Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v2.0-old.pl

* Scripts/webkitperl/LoadAsModule.pm: Added.
* Scripts/webkitperl/run-leaks_unittest/RunLeaks.pm: Removed.
(import):
(readFile):
* Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v1.0.pl: Used LoadAsModule instead of RunLeaks.
* Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v2.0-new.pl: Ditto.
* Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v2.0-old.pl: Ditto.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103338 268f45cc-cd09-0410-ab3c-d52691b4dbfc