IndexedDB: IDBKeyRange constructor should throw when lower > upper https://bugs.webkit.org/show_bug.cgi?id=74705
Reviewed by Tony Chang.

Source/WebCore:

Make IDBKeyRange throw an exception when lower > upper,
or lower == upper and one or both of the bounds is open.

Tested in storage/indexeddb/keyrange.html.

* storage/IDBKeyRange.cpp:
(WebCore::IDBKeyRange::bound):

LayoutTests:

Update tests to check that the IDBKeyRange constructor throws when
trying to specify illegal key ranges, and update tests that were
doing that.

* storage/indexeddb/index-cursor-expected.txt:
* storage/indexeddb/index-cursor.html:
* storage/indexeddb/keyrange-expected.txt:
* storage/indexeddb/keyrange.html:
* storage/indexeddb/objectstore-cursor-expected.txt:
* storage/indexeddb/objectstore-cursor.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103762 268f45cc-cd09-0410-ab3c-d52691b4dbfc