[chromium] LocalFileSystemChromium needs some thread safety fixes. https://bugs.webkit.org/show_bug.cgi?id=75494
Reviewed by Dmitry Titov.

This is essentially the same fix as r102894 except in slightly different code.

The important part of this fix is the removal of AllowCrossThreadAccess so
that ref counting happens appropriately.

Minor clean up throughout: Removed unnecessary WTF prefix in many
of these places and unnecessary String().

* src/LocalFileSystemChromium.cpp: See r102894 since this
is basically the same thing.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103996 268f45cc-cd09-0410-ab3c-d52691b4dbfc