[Chromium] Re-use TestShellLinux.cpp for Android https://bugs.webkit.org/show_bug.cgi?id=75169
Reviewed by Tony Chang.

TestShellLinux.cpp can be re-used for the test-shell on the WebKit
Chromium port for Android. Most custom inclusions for the DumpRenderTree
target are done in Tools/DumpRenderTree.gypi, but since *Linux.cpp files
are being excluded by a global rule on sources it has to be later in
the project generation process.

* WebKit.gyp:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104033 268f45cc-cd09-0410-ab3c-d52691b4dbfc