Valid canonical URLs should have a lowercase hostname https://bugs.webkit.org/show_bug.cgi?id=75771
Reviewed by Adam Barth.

Source/WebCore:

According to the RFC 3986 (and other browsers implementation), the hostname
of valid canonical URLs should be lowercase.

This patch lowercase the hostname in KURL::parse() similarily to what we
do for the scheme.

Tests: fast/url/host-lowercase-per-scheme.html
       fast/url/safari-extension.html

* platform/KURL.cpp:
(WebCore::isCanonicalHostnameLowercaseForScheme):
(WebCore::KURL::parse):

LayoutTests:

Update the test results for existing tests validating the hostname.

Added the test safari-extension.html for checking regressions of
the safari-extension:// protocol.

Added host-lowercase-per-scheme.html for testing the hostname is
set to lowercase for regognized Internet protocols.

* fast/url/host-expected.txt:
* fast/url/host-lowercase-per-scheme-expected.txt: Added.
* fast/url/host-lowercase-per-scheme.html: Added.
* fast/url/idna2008-expected.txt:
* fast/url/ipv4-expected.txt:
* fast/url/ipv6-expected.txt:
* fast/url/relative-expected.txt:
* fast/url/relative-unix-expected.txt:
* fast/url/safari-extension-expected.txt: Added.
* fast/url/safari-extension.html: Added.
* fast/url/standard-url-expected.txt:
* platform/chromium/fast/url/host-lowercase-per-scheme-expected.txt: Added.
* platform/chromium/fast/url/safari-extension-expected.txt: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104414 268f45cc-cd09-0410-ab3c-d52691b4dbfc