[Refactoring] Use join(", ", @arguments) to build a method argument string in CodeGeneratorV8.pm https://bugs.webkit.org/show_bug.cgi?id=75828
Reviewed by Darin Adler.

The code in CodeGeneratorV8.pm to build a method argument string is dirty
and error-prone. It is concatenating arguments one by one judging whether
", " is necessary or not. This patch refactors the code so that it pushes
all arguments into @arguments and then builds a method string by
join(", ", @arguments).

Test: bindings/scripts/test/*

* bindings/scripts/CodeGeneratorV8.pm:
(GenerateFunctionCallString):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104438 268f45cc-cd09-0410-ab3c-d52691b4dbfc