Improve keyboard navigation in layout test results https://bugs.webkit.org/show_bug.cgi?id=75901
Reviewed by Ojan Vafai.

Add support for keyboard shortcuts for navigating the list of tests,
and for flagging tests.

'i', 'k', 'k', and 'l' go foward, back, to first and to last, denoting
the current test with a red expand indicator.

'e', 'c' and 't' expand, collapse and toggle the current test.

'f' toggled a test as "flagged", which adds it to a copyable list
at the bottom of the document. The user can use flagging for whatever they want.

* fast/harness/resources/results-test.js:
(runTests):
(runTests.isCollapsed):
():
* fast/harness/results-expected.txt:
* fast/harness/results.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104606 268f45cc-cd09-0410-ab3c-d52691b4dbfc