Borders and box masks behave incorrectly with overlapping offsets https://bugs.webkit.org/show_bug.cgi?id=76137
Source/WebCore:

Reviewed by Dave Hyatt.

When border-image-slice sizes add up to more than the height or width
of the border-image, the middle sections should not be rendered, per spec.

Test: fast/borders/border-image-slice-constrained.html

* rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::paintNinePieceImage):

LayoutTests:

Reviewed by Dave Hyatt.

Testcase with border-image slices adding up to more than the height/width
of the border image.

Fixed cross-fade-webkit-mask-box-image.html to use slices that are half
of the image width.

* animations/cross-fade-webkit-mask-box-image.html:
* fast/borders/border-image-slice-constrained.html: Added.
* platform/mac/animations/cross-fade-webkit-mask-box-image-expected.png:
* platform/mac/fast/borders/border-image-slice-constrained-expected.png: Added.
* platform/mac/fast/borders/border-image-slice-constrained-expected.txt: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104851 268f45cc-cd09-0410-ab3c-d52691b4dbfc