Source/WebCore: Don't artifically keep IDBDatabase objects alive if there are no references to it. https://bugs.webkit.org/show_bug.cgi?id=75859
Originally, this code was added to keep the IDBDatabase object alive
even if no reference from JavaScript to the object existed, because
running transactions could still send events for this database
connection. Meanwhile, transactions are marked as active DOM objects
during their lifetime, and they keep a RefPtr to the IDBDatabase
object, so this hack is no longer required.

Reviewed by Tony Gentilcore.

Test: storage/indexeddb/delete-closed-database-object.html

* storage/IDBDatabase.cpp:
* storage/IDBDatabase.h:

LayoutTests: Test that an IDBDatabase object is garbage collected once no further references to it exist.
https://bugs.webkit.org/show_bug.cgi?id=75859

Reviewed by Tony Gentilcore.

* storage/indexeddb/delete-closed-database-object-expected.txt: Added.
* storage/indexeddb/delete-closed-database-object.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104933 268f45cc-cd09-0410-ab3c-d52691b4dbfc