prepare-ChangeLog: Look for a text editor in a few more variables. https://bugs.webkit.org/show_bug.cgi?id=76289
Reviewed by Ryosuke Niwa.

A text editor was looked for only in the CHANGE_LOG_EDITOR and
CHANGE_LOG_EDIT_APPLICATION environment variables before falling back
to the OS X-specific open(1) command.

As there does not seem to be a common way of looking for a text editor
among the scripts in Tools/Scripts right now, just look at $VISUAL and
$EDITOR (set much more often than the other variables in Unix systems)
before falling back to open(1).

* Scripts/prepare-ChangeLog:
(openChangeLogs):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104955 268f45cc-cd09-0410-ab3c-d52691b4dbfc