test-webkitpy: add basic code coverage support https://bugs.webkit.org/show_bug.cgi?id=76201
Reviewed by Adam Barth.

This test adds basic code coverage support to test-webkitpy; if
you pass the -c flag, we will generate a new coverage file. We
can start simply with this and add more features as necessary.

This patch requires coverage to be already installed on the
machine. I am not auto-installing this for now.

* Scripts/webkitpy/test/main.py:
(Tester.parse_args):
(Tester.run_tests):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104975 268f45cc-cd09-0410-ab3c-d52691b4dbfc