WebCore build exceeds address space on 32-bit Windows builders (again). https://bugs.webkit.org/show_bug.cgi?id=77357
Reviewed by Adam Roben.

Add an Inspector All-In-One file, but only use it in Release and Production builds.
This differs from our other All-In-One files, but it is a better approach because it makes
debugging possible in the Debug configuration (the symbols will be in the correct .obj file
for the original .cpp files).

* WebCore.vcproj/WebCore.vcproj:
Also, let VS have its way with the vcproj file.
* inspector/InspectorAllInOne.cpp: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106280 268f45cc-cd09-0410-ab3c-d52691b4dbfc