ASSERT(m_jumpsToLink.isEmpty()) failing in ARMv7Assembler dtor https://bugs.webkit.org/show_bug.cgi?id=77443
Reviewed by Gavin Barraclough.

Removed failing ASSERT() and thus destructor.  The ASSERT isn't needed.
We are hitting it in the YARR JIT case where we bail out and go to the
interpreter with a partially JIT'ed function.  Since we haven't linked
the JIT'ed code, there is likely to be some unresolved jumps in the vector
when the ARMv7Assembler destructor is called.  For the case where we
complete the JIT process, we clear the vector at the end of
LinkBuffer::linkCode (LinkBuffer.h:292).

* assembler/ARMv7Assembler.h:
(ARMv7Assembler):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106370 268f45cc-cd09-0410-ab3c-d52691b4dbfc