Hidden form elements do not save their state prior to form submission https://bugs.webkit.org/show_bug.cgi?id=77391 <rdar://problem/10563108>
Reviewed by Brady Eidson.

Source/WebCore:

Test: fast/forms/state-restore-hidden.html

* html/HiddenInputType.cpp: Teach hidden inputs to save and restore their state.
(WebCore::HiddenInputType::saveFormControlState):
(WebCore::HiddenInputType::restoreFormControlState):
* html/HiddenInputType.h:
(HiddenInputType):

LayoutTests:

The test changes the value of the hidden input element via Javascript prior to
submitting the form. The form action goes back one page, and we check to see
that the hidden input value is what it was set to by Javascript, rather than what
it was originally assigned.

* fast/forms/state-restore-hidden-expected.txt: Added.
* fast/forms/state-restore-hidden.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106388 268f45cc-cd09-0410-ab3c-d52691b4dbfc