Crash in Node::dispatchSubtreeModifiedEvent https://bugs.webkit.org/show_bug.cgi?id=77449
Reviewed by Alexey Proskuryakov.

Source/WebCore: 

The bug was caused by appendChild not retaining this pointer.
This is normally okay because there's another owner within JSC/V8 binding code that
holds onto the node but this isn't the case when nodes are created as a part
of setting document.title. Fixed the crash by retaining the pointer as needed.

Test: fast/dom/remove-body-during-title-creation.html

* dom/ContainerNode.cpp:
(WebCore::ContainerNode::appendChild):

LayoutTests: 

Add a regression test for the crash.

* fast/dom/remove-body-during-title-creation-expected.txt: Added.
* fast/dom/remove-body-during-title-creation.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106715 268f45cc-cd09-0410-ab3c-d52691b4dbfc