Layout Test http/tests/security/xssAuditor/dom-write-innerHTML.html is flaky https://bugs.webkit.org/show_bug.cgi?id=67538
Reviewed by Kentaro Hara.

Try to fix the flakiness in this test by kicking off the <img> load
before the load event has fired for the inner frame.  That way, the
load event for the main frame will be delayed until the image finishes
loading (and its error handler has fired), removing the need for the
setTimeout.

* http/tests/security/xssAuditor/dom-write-innerHTML.html:
* http/tests/security/xssAuditor/resources/echo-dom-write-innerHTML.html:
* platform/chromium/test_expectations.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106785 268f45cc-cd09-0410-ab3c-d52691b4dbfc