Remove [ConvertingNullStringTo] from CloseEvent.idl https://bugs.webkit.org/show_bug.cgi?id=78328
Reviewed by Adam Barth.

In CloseEvent.idl, [ConvertingNullStringTo] is a typo of [ConvertNullStringTo],
(although in bug 78108, [ConvertNullStringTo] was renamed to [TreatReturnedNullStringAs]).

Anyway, the spec says that "The reason attribute must return the value it was
initialized to. When the object is created, this attribute must be initialized to empty string."
http://dev.w3.org/html5/websockets/#event-definitions
Thus, this patch removes [ConvertingNullStringTo] from CloseEvent.idl.

Test: fast/events/constructors/close-event-constructor.html

* websockets/CloseEvent.idl:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107382 268f45cc-cd09-0410-ab3c-d52691b4dbfc