Add [JSCustomToJSObject] IDL attribute to interfaces that have custom toJS() but do not have custom toV8() https://bugs.webkit.org/show_bug.cgi?id=78466
Reviewed by Adam Barth.

This is the second step to remove hard-coding in HasCustomToV8Implementation()
in CodeGeneratorV8.pm. This patch replaces [JSCustomToJS] with [JSCustomToJSObject]
for interfaces which have custom toJS() but do not have custom toV8().

No tests. No change in behavior.

* bindings/scripts/CodeGeneratorJS.pm:
(GenerateHeader):
(GenerateImplementation):
* bindings/scripts/CodeGeneratorV8.pm:
(HasCustomToV8Implementation): I found that AbstractWorker and CanvasRenderingContext
are the only IDL files to which I need to add [JSCustomToJSObject].
Other IDL files which had been listed here do not have [JSCustomToJS].
* html/canvas/CanvasRenderingContext.idl:
* workers/AbstractWorker.idl:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107553 268f45cc-cd09-0410-ab3c-d52691b4dbfc