StylePropertySet: Try to find a CSSValuePool for identifier values. <http://webkit.org/b/78590>
Reviewed by Antti Koivisto.

For StylePropertySets with a parent element (inline and attribute style),
grab at the parentElement()->document() to locate a CSSValuePool.
This code will change soon in the CSSOM refactoring and moreso once we
figure out how to have a global CSSValuePool, but I'm hoping it will buy
us some perf back in the meantime.

* css/StylePropertySet.cpp:
(WebCore::StylePropertySet::setProperty):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107699 268f45cc-cd09-0410-ab3c-d52691b4dbfc