NPN_GetValueForURL / NPNURLVProxy returns DIRECT when proxy configured via PAC <http://webkit.org/b/78766> / <rdar://problem/10729283>
Reviewed by Anders Carlsson.

* platform/network/cf/ProxyServerCFNet.cpp:
(WebCore::proxyAutoConfigurationResultCallback): Stop the runloop, and then process
the results that we received.
(WebCore::processProxyServers): Processing of array of proxy configuration information
moved from addProxyServersForURL. Handling of proxy auto-configuration URLs is now handled
by calling CFNetworkExecuteProxyAutoConfigurationURL and waiting synchronously on the result
callback. Doing this synchronously is not great, but it's the best we can do without a lot
of restructuring of the code that calls this. We arbitrarily time out the execution after five
seconds to avoid permanently hanging.
(WebCore::addProxyServersForURL): Call in to our helper function.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107966 268f45cc-cd09-0410-ab3c-d52691b4dbfc