[chromium] Fix image drag out on Chromium https://bugs.webkit.org/show_bug.cgi?id=79158
Reviewed by Tony Chang.

Source/WebKit/chromium:

We constructed the WebDragData::Item for dragging out an image but never added it to the
item list. This is normally covered by the fast/events/drag-image-filename.html layout test,
but the bug wasn't caught in the original patch since the implementations in Chrome and DRT
had diverged.

* src/WebDragData.cpp:
(WebKit::WebDragData::items):

Tools:

* DumpRenderTree/chromium/EventSender.cpp:
(EventSender::dumpFilenameBeingDragged):
(EventSender::beginDragWithFiles):
* DumpRenderTree/chromium/WebViewHost.cpp:
(addDRTFakeFileToDataObject):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@108419 268f45cc-cd09-0410-ab3c-d52691b4dbfc