[chromium] Have WebFrameImpl::selectionAsMarkup create interchange markup. https://bugs.webkit.org/show_bug.cgi?id=77307
Patch by Peter Collingbourne <peter@pcc.me.uk> on 2012-02-24
Reviewed by Tony Chang.

This function is currently used to build markup in order to
print the selection, and will also be used to build markup for
the clipboard.  Without this change, the markup will refer to
(for example) nonexistent CSS classes and invalid relative
paths and therefore will not be interpreted correctly.

Chromium bug: http://crbug.com/23329

* src/WebFrameImpl.cpp:
(WebKit::WebFrameImpl::selectionAsMarkup):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@108821 268f45cc-cd09-0410-ab3c-d52691b4dbfc