JSString::resolveRope() should report extra memory cost to the heap. https://bugs.webkit.org/show_bug.cgi?id=79555
Patch by Yong Li <yoli@rim.com> on 2012-02-28
Reviewed by Michael Saboff.

At the time a JSString is constructed with fibers, it doesn't report
extra memory cost, which is reasonable because it hasn't allocate
new memory. However when the rope is resolved, it should report meory
cost for the new buffer.

* runtime/JSString.cpp:
(JSC::JSString::resolveRope):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@109105 268f45cc-cd09-0410-ab3c-d52691b4dbfc