<msubsup> setNeedsLayout() correction https://bugs.webkit.org/show_bug.cgi?id=79856
Patch by David Barton <dbarton@mathscribe.com> on 2012-02-29
Reviewed by Julien Chaffraix.

m_scripts->setNeedsLayout(true, false); needs to be added to the fix for bug 79274.

No new tests. It's actually difficult to test the need for this with the current code,
because there are a lot of extra setNeedsLayout() calls right now. I caught this after
removing some of those calls that should be redundant, for a future patch. Actually
removing those calls now changes other behavior, due to other bugs, so I'll save all
that for future patches. We should land this fix now though I think, as it could
possibly matter in some cases.

* rendering/mathml/RenderMathMLSubSup.cpp:
(WebCore::RenderMathMLSubSup::layout):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@109286 268f45cc-cd09-0410-ab3c-d52691b4dbfc