[Qt] Use config test to determine if the compiler supports -MD
We used to assume that unix systems building against Qt5 would have
a compiler that supported -MD. Although this was usually the case,
we failed to take into account broken tools for distributed builds
(e.g. teambuilder or icecream) which would end up producing .d files
in the root of the build dir instead of along-side the object file.

Now that we use a config-test to determine the support, we'll capture
those cases, unless distributed builds are enabled after the initial
build-webkit run.

Reviewed by Csaba Osztrogonác.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@109733 268f45cc-cd09-0410-ab3c-d52691b4dbfc