Element not fully repainted after application and removal of transform https://bugs.webkit.org/show_bug.cgi?id=61338
Source/WebCore:

Patch by Kishore Bolisetty <kbolisetty@innominds.com> on 2012-03-05
Reviewed by Simon Fraser.
RenderLayer is created and removed when a transform is applied and removed to element.
To render the element properly after removal of transform on it, preferred widths and dimensions of
the element and its containing block needs to be recalculated.

Test: fast/repaint/transform-rotate-and-remove.html

* rendering/RenderBoxModelObject.cpp:
(WebCore):
(WebCore::RenderBoxModelObject::styleWillChange):
(WebCore::RenderBoxModelObject::styleDidChange):
* rendering/RenderBoxModelObject.h:
(RenderBoxModelObject):

LayoutTests:

Patch by Kishore Bolisetty <kbolisetty@innominds.com> on 2012-03-05
Reviewed by Simon Fraser.
Test to remove the transform property of an element upon onload event.

* fast/repaint/transform-rotate-and-remove.html: Added.
* platform/chromium-linux/fast/repaint/transform-rotate-and-remove-expected.png: Added.
* platform/chromium-linux/fast/repaint/transform-rotate-and-remove-expected.txt: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@109867 268f45cc-cd09-0410-ab3c-d52691b4dbfc