[Qt][WK2] Add support for rudimentary scroll indicators in MiniBrowser https://bugs.webkit.org/show_bug.cgi?id=80832
Reviewed by Tor Arne Vestbø.

Source/WebKit2:

Since the ScrollDecorator QML component requires a Flickable in its API
we need to expose the Flickable in QML for now and we also need to add
a notifier because we instantiate the internal Flickable only when the
WebView component completes construction.

* UIProcess/API/qt/qquickwebview.cpp:
(QQuickWebViewFlickablePrivate::onComponentComplete):
(QQuickWebViewExperimental::flickable):
* UIProcess/API/qt/qquickwebview_p.h:

Tools:

Add a basic scroll indicator component to MiniBrowser so that it
behaves similar to the ScrollDecorator QML component.

* MiniBrowser/qt/MiniBrowser.qrc:
* MiniBrowser/qt/qml/BrowserWindow.qml:
* MiniBrowser/qt/qml/ScrollIndicator.qml: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110447 268f45cc-cd09-0410-ab3c-d52691b4dbfc