[BlackBerry] Remove sublayers before clamping just in case the layer already exists https://bugs.webkit.org/show_bug.cgi?id=80989
Patch by Konrad Piascik <kpiascik@rim.com> on 2012-03-13
Reviewed by Antonio Gomes.

Discovered when trying to add a highlight sublayer for web inspector.  Since that layer
may already exist in the tree we need to see if it will be pruned before we clamp on the
size.

This was manually tested during the writing of another patch to add web inspector highlights
to accelrated compositing.

* platform/graphics/blackberry/LayerWebKitThread.cpp:
(WebCore::LayerWebKitThread::insertSublayer):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110578 268f45cc-cd09-0410-ab3c-d52691b4dbfc