[BlackBerry] Upstream the BlackBerry change to platform/Cursor.h https://bugs.webkit.org/show_bug.cgi?id=81086
Reviewed by Rob Buis.

Source/WebCore:

* platform/Cursor.h:
(WebCore):
(WebCore::Cursor::Cursor):
Add the BlackBerry specific PlatformCursor.
* platform/blackberry/PageClientBlackBerry.h:
There was an alias named PlatformCursorHandle for PlatformCursor
in the internal version of Cursor.h. But it's not necessary so it's
not included in the upstreaming version of Cursor.h. As a result,
use PlatformCursor directly in this file.
(PageClientBlackBerry):

Source/WebKit/blackberry:

There was an alias named PlatformCursorHandle for PlatformCursor
in the internal version of Cursor.h. But it's not necessary so it's
not included in the upstreaming version of Cursor.h. As a result,
use PlatformCursor directly.

* Api/WebPage.cpp:
(BlackBerry::WebKit::WebPagePrivate::setCursor):
* Api/WebPage_p.h:
(WebPagePrivate):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110701 268f45cc-cd09-0410-ab3c-d52691b4dbfc