Evict tiles from pages in background tabs https://bugs.webkit.org/show_bug.cgi?id=81829 <rdar://problem/10866152>
Reviewed by Andreas Kling.

When the tile cache for a page is no longer in a window (which happens when it's moved to
a background tab), schedule a tile revalidation after 4 seconds. This tile revalidation
will ensure that tiles outside of the visible rect will be dropped.

* platform/graphics/ca/mac/TileCache.h:
(TileCache):
* platform/graphics/ca/mac/TileCache.mm:
(WebCore::TileCache::tileCacheLayerBoundsChanged):
(WebCore::TileCache::setIsInWindow):
(WebCore::TileCache::tileCoverageRect):
(WebCore):
(WebCore::TileCache::scheduleTileRevalidation):
(WebCore::TileCache::revalidateTiles):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111604 268f45cc-cd09-0410-ab3c-d52691b4dbfc