WebKitURLWithTitles pasteboard format should support URLs containing Emoji characters. https://bugs.webkit.org/show_bug.cgi?id=81835 <rdar://problem/11082749>
Reviewed by Brady Eidson.

Source/WebCore: 

When writing the URL in the pasteboard for the WebURLWithTitles format,
we use the user visible string instead of the original URL.
In case of URL with Unicode characters, we lose the encoding.
        
Added TestWebKitAPI test.

* platform/mac/PasteboardMac.mm:
(WebCore::writeURLForTypes): Writes to the pasteboard the original URL.

Tools: 

* TestWebKitAPI/Tests/mac/ContextMenuCanCopyURL.html:
* TestWebKitAPI/Tests/mac/ContextMenuCanCopyURL.mm:
(TestWebKitAPI::contextMenuCopyLink):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111614 268f45cc-cd09-0410-ab3c-d52691b4dbfc