[chromium] Target surface should be damaged for a new layers even when layer had no changes https://bugs.webkit.org/show_bug.cgi?id=81879
Reviewed by Adrienne Walker.

Source/WebCore:

Unit test added to CCDamageTrackerTest.

* platform/graphics/chromium/cc/CCDamageTracker.cpp:
(WebCore::CCDamageTracker::removeRectFromCurrentFrame): added a
boolean arg to detect if the layer is new on this update.

(WebCore::CCDamageTracker::extendDamageForLayer): added logic that
damages the target surface if the layer is new.

(WebCore::CCDamageTracker::extendDamageForRenderSurface): added
logic that damages the target surface if the descendant surface is
new; similar logic for the surface's replica if the replica is new.

* platform/graphics/chromium/cc/CCDamageTracker.h:
(CCDamageTracker):

Source/WebKit/chromium:

* tests/CCDamageTrackerTest.cpp:
(WebKitTests::TEST_F):
(WebKitTests):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111817 268f45cc-cd09-0410-ab3c-d52691b4dbfc