Unreviewed buildfix for the Chromium Android builder.
Disable the ImageDiff target for now. Android needs to build this for
host rather than for target (as other configurations do), but since the
WTF move the dependency chain got significantly more complicated.

I'll work out a proper fix in the following bug, but since several
things are broken due to this, this buildfix should be landed first.
https://bugs.webkit.org/show_bug.cgi?id=82039

* DumpRenderTree/DumpRenderTree.gyp/DumpRenderTree.gyp:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111846 268f45cc-cd09-0410-ab3c-d52691b4dbfc