REGRESSION(107971): Google Voice contact list is broken in WebKit due to badly allocating the extra height https://bugs.webkit.org/show_bug.cgi?id=81826
Reviewed by Tony Chang.

Source/WebCore:

Covered by tables/mozilla/bugs/bug27038-{1|2}.html.

This partly reverts r107971: the extra logical height distribution change was not needed
to fix the bug (it is needed by the test though). We revert to giving all the extra height
to the first tbody and not the first section.

This is broken but unfortunately some websites are relying on that. Getting a real
distribution algorithm is covered by bug 81824. However this is super tricky to get
right and I did not want to add more compatibility risks until I have something solid.

* rendering/RenderTable.cpp:
(WebCore::RenderTable::distributeExtraLogicalHeight):

LayoutTests:

* fast/table/double-height-table-no-tbody.html-disabled: Renamed from LayoutTests/fast/table/double-height-table-no-tbody.html.
Disabled this test as it relies on our algorithm not to discriminate between first sections.

* platform/chromium/test_expectations.txt:
* platform/qt/Skipped:
Mark those 2 tests as needing a new baseline again.

* platform/efl/test_expectations.txt:
* platform/gtk/Skipped:
* platform/mac/test_expectations.txt:
* platform/win/Skipped:
Those platforms did not rebaseline those 2 tests so they should automatically pass them.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111882 268f45cc-cd09-0410-ab3c-d52691b4dbfc