[GTK] libgcrypt and p11-kit should not be in jhbuild modules https://bugs.webkit.org/show_bug.cgi?id=82073
Reviewed by Martin Robinson.

* Scripts/webkitdirs.pm:
(jhbuildConfigurationChanged): new function that isolates the
jhbuild configuration change checking from the autogen logic.
(mustReRunAutogen): code for the new function lived here.
(buildAutotoolsProject): now calls jhbuildConfigurationChanged
itself, and cleans up the jhbuild root if configuration has
changed; it also forces autogen to be run in that case, to
make sure the new libraries are used.
* gtk/jhbuild.modules: removed libgcrypt and p11-kit.
* jhbuild/jhbuild-wrapper:
(ensure_jhbuild): this might be run inside a jhbuild environment,
in which case aclocal fails when trying to use the now deleted
aclocal directory in the jhbuild prefix, so work around that.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111929 268f45cc-cd09-0410-ab3c-d52691b4dbfc