Fix flaky test fast/eventsource/eventsource-url-attribute.html https://bugs.webkit.org/show_bug.cgi?id=82257
Patch by Pablo Flouret <pablof@motorola.com> on 2012-03-26
Reviewed by Ryosuke Niwa.

The test used http://localhost, chosen as a semi-random url, but since
it doesn't return a text/event-stream mime-type sometimes a console
warning would appear. Moved the test to the http folder and changed it
to use a resource with the correct mime-type.

* http/tests/eventsource/eventsource-url-attribute-expected.txt: Renamed from LayoutTests/fast/eventsource/eventsource-url-attribute-expected.txt.
* http/tests/eventsource/eventsource-url-attribute.html: Renamed from LayoutTests/fast/eventsource/eventsource-url-attribute.html.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112204 268f45cc-cd09-0410-ab3c-d52691b4dbfc