cssText should not generate literal 'initial' in shorthand properties https://bugs.webkit.org/show_bug.cgi?id=82364
Reviewed by Antti Koivisto.

Source/WebCore: 

Fixed the bug by treating initial value as if the value is not set.
While this is incorrect for properties that inherits by default,
it's strictly better than generating unparsable value as we do today.

The proper fix is for CSSInitialValue::cssText to fetch the respective
default value from what's currently in CSSStyleSelector code but that requires
a considerable amount of refactoring and work.

* css/StylePropertySet.cpp:
(WebCore::StylePropertySet::getShorthandValue):
(WebCore::StylePropertySet::getCommonValue):

LayoutTests: 

* fast/css/cssText-shorthand-expected.txt: Rebaselined.
* fast/css/cssText-shorthand.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112321 268f45cc-cd09-0410-ab3c-d52691b4dbfc