webkit_unit_tests crashing on chromium lion in CCLayerTreeHostTestCompositeAndReadbackWhileInvisible https://bugs.webkit.org/show_bug.cgi?id=82382
Patch by James Robinson <jamesr@chromium.org> on 2012-03-27
Reviewed by Adrienne Walker.

Use UnsignedWithZeroKeyHashTraits trait for the HashSet<WebGLId> m_usedTextures in
CompositorFakeWebGraphicsContext3DWithTextureTracking. This set is used to track all textures that are bound in
a given context during a test. During ..CompositeAndReadback.. we bind texture id 0. By default, HashSet<T>
where T is an unsigned does not allow 0 as a value. WebGLId is typedef'd to unsigned.

* tests/CCLayerTreeHostTest.cpp:
(CompositorFakeWebGraphicsContext3DWithTextureTracking):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112356 268f45cc-cd09-0410-ab3c-d52691b4dbfc