REGRESSION (r109826): Can't type into Flash text fields https://bugs.webkit.org/show_bug.cgi?id=82488 <rdar://problem/11022004>
Reviewed by Sam Weinig.

This was caused by the fix in r109826. Revert that change and fix the original bug by notifying the UI process
that the plug-in lost focus when it's destroyed. This will correctly reset the text input state (merely setting
the text input state to PluginComplexTextInputDisabled doesn't reset the state correctly).

* UIProcess/API/mac/WKView.mm:
(-[WKView _setPluginComplexTextInputState:]):
(-[WKView _handlePluginComplexTextInputKeyDown:]):
* WebProcess/Plugins/PluginView.cpp:
(WebKit::PluginView::~PluginView):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112415 268f45cc-cd09-0410-ab3c-d52691b4dbfc