Respects font fallback list during webfonts are loading https://bugs.webkit.org/show_bug.cgi?id=76684
Reviewed by Dimitri Glazkov.

Source/WebCore:

For layout, use the rest of the fallback list while webfonts are loading.
If the webfont is loading, it will be given an invalid Unicode Range so that
font selection mechansim doesn't use the webfont for layout.

Test: http/tests/webfont/fallback-font-while-loading.html

* css/CSSSegmentedFontFace.cpp:
(WebCore::appendFontDataWithInvalidUnicodeRangeIfLoading): Added.
(WebCore):
(WebCore::CSSSegmentedFontFace::getFontData):

LayoutTests:

Add a test that checks fallback fonts are used for layout while
webfonts are loading.

* http/tests/webfont/fallback-font-while-loading-expected.txt: Added.
* http/tests/webfont/fallback-font-while-loading.html: Added.
* http/tests/webfont/getahem.cgi: Added.
* http/tests/webfont/resources/Ahem.ttf: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112489 268f45cc-cd09-0410-ab3c-d52691b4dbfc