[Qt] WK1/WK2 layout test results are diverging https://bugs.webkit.org/show_bug.cgi?id=82720
The resolving of default font families was altered in Qt5.
This directly affects the font picked by DRT for layout
tests. Since we want to keep the Qt4 and Qt5 results in line,
we can trick DRT in picking the same font as before for now.

Patch by Pierre Rossi <pierre.rossi@gmail.com> on 2012-04-05
Reviewed by Csaba Osztrogonác.

* DumpRenderTree/qt/QtInitializeTestFonts.cpp:
(WebKit::initializeTestFonts):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113321 268f45cc-cd09-0410-ab3c-d52691b4dbfc