[Gtk] Web Inspector noinst_DATA images are copied into innacurately named directory https://bugs.webkit.org/show_bug.cgi?id=83423
Patch by Zan Dobersek <zandobersek@gmail.com> on 2012-04-09
Reviewed by Martin Robinson.

Copy Web Inspector images that are a part of the data not meant
for installation into a directory named 'Images' rather than
a lower-case version of that. This is required as until now,
when using these inspector resources (for example during layout
tests or manually pointing WEBKIT_INSPECTOR_PATH env to that
location), the images were not displayed as they were not loadable.

No new tests - no new functionality.

* GNUmakefile.am:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113560 268f45cc-cd09-0410-ab3c-d52691b4dbfc