Correct LayoutUnit usage in style functions getRoundedBorderFor https://bugs.webkit.org/show_bug.cgi?id=83154
Reviewed by Julien Chaffraix.

Correcting LayoutUnit usage in getRoundedBorderFor and the calcRadiiFor helper function.

No new tests. No change in behavior.

* rendering/style/RenderStyle.cpp:
(WebCore::calcRadiiFor): This is a static internal function that returns the radii based on a size.
It's only called from getRoundedBorderFor, so we're now pixel snapping the size before handing it
to this function.
(WebCore::RenderStyle::getRoundedBorderFor): Takes a LayoutRect and returns the pixel-snapped
resulting RoundedRect. Using a local IntRect to avoid snapping twice.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113594 268f45cc-cd09-0410-ab3c-d52691b4dbfc