Fix the build with gcc 4.7.0 https://bugs.webkit.org/show_bug.cgi?id=83584
Reviewed by Alexey Proskuryakov.

Source/WebCore:

No new tests needed.

* Target.pri: Disable Werror on the ANGLE code since it has several recurring issues
because older versions of flex (<2.5.35) and bison (<2.5) generate code which
triggers warnings. This generated code is committed on each version update of ANGLE,
and it is very likely that the code is generated with older versions of the mentioned
tools.
* html/HTMLImageElement.cpp:
(WebCore::HTMLImageElement::createForJSConstructor): Fails because of -Werror=extra
* platform/graphics/TiledBackingStoreClient.h: Fails because -Werror=delete-non-virtual-dtor,
gcc requires classes which have virtual methods to have a virtual destructor, if their instances
or the instances of their subclasses are deleted.
(WebCore::TiledBackingStoreClient::~TiledBackingStoreClient):
* platform/graphics/texmap/TextureMapperPlatformLayer.h: Ditto.
(WebCore::TextureMapperPlatformLayer::~TextureMapperPlatformLayer):

Tools:

* qmake/mkspecs/features/unix/default_post.prf:
The -fuse-ld=gold option is not supported with upstream gcc.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113848 268f45cc-cd09-0410-ab3c-d52691b4dbfc