ThreadingWin: Mutex::unlock() can be "over-unlocked". https://bugs.webkit.org/show_bug.cgi?id=83725
Reviewed by David Levin.

In order to support the behavior of pthread_mutex_trylock(), the Windows Mutex class includes
a recursion counter which is incremented in Mutex::lock(), decremented in Mutex::unlock(),
and checked in Mutex::tryLock().  If the mutex is "over-unlocked", the counter wraps around to
MAX_INT, and subsequent calls to Mutex::trylock() will fail. Raise an ASSERT in this situation
so the "over-unlock" will be caught.

* wtf/ThreadingWin.cpp:
(WTF::Mutex::unlock): ASSERT if unlocking a non-locked mutex.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114028 268f45cc-cd09-0410-ab3c-d52691b4dbfc