CSSValuePool: Made identifier value cache a fixed-size array. <http://webkit.org/b/84219>
Reviewed by Antti Koivisto.

Change the identifier CSSPrimitiveValue cache in CSSValuePool from a HashMap to a
fixed-size array of RefPtr<CSSPrimitiveValue>s.

We have ~700 values total, so this is quite space efficient now that the CSSValuePool
is globally shared. More importantly it avoids a hash lookup every time we need an
identifier value.

* css/CSSValuePool.h:
* css/CSSValuePool.cpp:
(WebCore::CSSValuePool::createIdentifierValue):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114474 268f45cc-cd09-0410-ab3c-d52691b4dbfc