DevTools highlights elements in frames at un-scaled positions https://bugs.webkit.org/show_bug.cgi?id=84181
Source/WebCore:

On mobile platforms (e.g. chromium-android), normally pages are scaled,
thus the coordinations of highlight rect inside of a frame needs to
consider not only the offset, but also the scale. The change uses
FrameView::contentsToRootView() to map the coordinates of a node in
a frame to the coordinates in the main frame, instead of calculating
by only the offset.

Reviewed by Pavel Feldman.

Test: inspector/elements/highlight-node-scaled.html

* inspector/DOMNodeHighlighter.cpp:

LayoutTests:

The new test case is a scaled web page containing an iframe. It tests
if the highlight rect of an element in the iframe is correct.

Reviewed by Pavel Feldman.

* inspector/elements/highlight-node-scaled-expected.txt: Added.
* inspector/elements/highlight-node-scaled.html: Added.
* inspector/elements/resources/highlight-node-scaled-iframe.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114659 268f45cc-cd09-0410-ab3c-d52691b4dbfc