Web Inspector: make ParsedURL.prototype.displayName data url friendly. https://bugs.webkit.org/show_bug.cgi?id=84578
Reviewed by Yury Semikhatsky.

We'd like to define valid URL as the one that has scheme, host and path.
We can append to this URL and manipulate its content. We still want
possibly invalid specs (such as data or about) to exist and have nice
display names.

* inspector/front-end/ResourceUtils.js:
(WebInspector.ParsedURL):
(WebInspector.ParsedURL.prototype.get displayName):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114882 268f45cc-cd09-0410-ab3c-d52691b4dbfc