Extra display logic for the media control panel element https://bugs.webkit.org/show_bug.cgi?id=82476
Patch by Victor Carbune <vcarbune@adobe.com> on 2012-04-24
Reviewed by Eric Carlson.

Source/WebCore:

This patch fixes a bug which caused the controls to be displayed
when they should remain hidden. Added an extra variable to the
panel elements which properly keeps the state of the panel (visible or not).

Test: media/video-controls-toggling.html

* html/shadow/MediaControlElements.cpp:
(WebCore::MediaControlPanelElement::MediaControlPanelElement): Added the
variable m_isDisplayed to hold the state whether the panel is visible or not.
(WebCore::MediaControlPanelElement::makeOpaque): Showing the panel only if it
is visible.
(WebCore::MediaControlPanelElement::makeTransparent): Enabled the transition
timer which sets the display:none property on the controls.
(WebCore::MediaControlPanelElement::setIsDisplayed): Setter for the state variable.
(WebCore):
* html/shadow/MediaControlElements.h:
(MediaControlPanelElement):
* html/shadow/MediaControlRootElement.cpp:
(WebCore::MediaControlRootElement::show): Updated the panel visibility state.
(WebCore::MediaControlRootElement::hide): Updated the panel visibility state.
* html/shadow/MediaControlRootElementChromium.cpp:
(WebCore::MediaControlRootElementChromium::show): Updated the panel visibility state.
(WebCore::MediaControlRootElementChromium::hide): Updated the panel visibility state.

LayoutTests:

Added test to ensure that controls are not displayed when
the controls attribute is not set.

* media/video-controls-toggling-expected.txt: Added.
* media/video-controls-toggling.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115125 268f45cc-cd09-0410-ab3c-d52691b4dbfc