Crash when collecting svg symbol element in named flow. https://bugs.webkit.org/show_bug.cgi?id=84493
Reviewed by David Hyatt.

Source/WebCore:

Test: fast/regions/symbol-in-named-flow-crash.svg

* dom/Element.cpp:
(WebCore::Element::~Element):
Add an assert that an element that was collected into a named flow was already removed at this point
(when the document is not in the process of destruction)
(WebCore::Element::unregisterNamedFlowContentNode):
Created a new function for unregistering a content node. In the future, this function may be used for
content nodes from shadow dom.
(WebCore::Element::detach):
* dom/Element.h:
(Element):
* dom/NodeRenderingContext.cpp:
(WebCore::NodeRenderingContext::moveToFlowThreadIfNeeded):
Prevent elements that are part of shadow dom to be collected into a named flow.

LayoutTests:

* fast/regions/symbol-in-named-flow-crash-expected.txt: Added.
* fast/regions/symbol-in-named-flow-crash.svg: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115297 268f45cc-cd09-0410-ab3c-d52691b4dbfc